LayoutTests/imported/w3c/web-platform-tests/svg/animations/dependent-begin-on-syncbase-expected.txt


FAIL Dependent timed element begin (restart) on syncbase assert_equals: expected 1 but got 0
